+++ packages/ckb-sdk-core/examples/sendSimpleTransaction.js
@@ -1,12 +1,25 @@
 /* eslint-disable */
+const path = require('path')
+const os = require('os')
+const { Indexer, CellCollector } = require('@ckb-lumos/indexer')
 const CKB = require('../lib').default
 
-const bootstrap = async () => {
-  const nodeUrl = process.env.NODE_URL || 'http://localhost:8114' // example node url
-  const privateKey = process.env.PRIV_KEY || '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee' // example private key
-  const blockAssemblerCodeHash = '0x1892ea40d82b53c678ff88312450bbb17e164d7a3e0a90941aa58839f56f8df2' // transcribe the block_assembler.code_hash in the ckb.toml from the ckb project
+const CKB_URL = process.env.CKB_URL || 'http://localhost:8114' // example node url
+const LUMOS_DB = path.join(os.tmpdir(), 'lumos_db')
+/**
+ * lumos indexer
+ */
+const indexer = new Indexer(CKB_URL, LUMOS_DB)
+indexer.startForever()
 
-  const ckb = new CKB(nodeUrl) // instantiate the JS SDK with provided node url
+/**
+ * sdk
+ */
+
+const bootstrap = async () => {
+  const privateKey = process.env.PRIV_KEY || '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee' // example private key
+
+  const ckb = new CKB(CKB_URL) // instantiate the JS SDK with provided node url
 
   await ckb.loadDeps() // load the dependencies of secp256k1 algorithm which is used to verify the signature in transaction's witnesses.
 
@@ -36,33 +49,15 @@ const bootstrap = async () => {
    */
   // console.log(JSON.stringify(addresses, null, 2))
 
-  /**
-   * calculate the lockHash by the address publicKeyHash
-   * 1. the publicKeyHash of the address is required in the args field of lock script
-   * 2. compose the lock script with the code hash(as a miner, we use blockAssemblerCodeHash here), and args
-   * 3. calculate the hash of lock script via ckb.utils.scriptToHash method
-   */
-
-  const lockScript = {
-    hashType: "type",
-    codeHash: blockAssemblerCodeHash,
-    args: publicKeyHash,
+  const lock = {
+    codeHash: ckb.config.secp256k1Dep.codeHash,
+    hashType: ckb.config.secp256k1Dep.hashType,
+    args: publicKeyHash
   }
   /**
-   * to see the lock script
+   * load cells from lumos as `examples/sendTransactionWithLumosCollector.js` shows
    */
-  // console.log(JSON.stringify(lockScript, null, 2))
-
-  const lockHash = ckb.utils.scriptToHash(lockScript)
-  /**
-   * to see the lock hash
-   */
-  // console.log(lockHash)
-
-  // method to fetch all unspent cells by lock hash
-  const unspentCells = await ckb.loadCells({
-    lockHash
-  })
+  const unspentCells = await ckb.loadCells({ indexer, CellCollector, lock })
 
   /**
    * to see the unspent cells
@@ -76,6 +71,13 @@ const bootstrap = async () => {
     prefix: 'ckt'
   })
 
+  /**
+   * @param fee - transaction fee, can be set in number directly, or use an reconciler to set by SDK
+   *                               say, fee: BigInt(100000) means transaction fee is 100000 shannons
+   *                                or, fee: { feeRate: '0x7d0', reconciler: ckb.utils.reconcilers.extraInputs } to set transaction fee by reconcilers.extraInputs with feeRate = 2000 shannons/Byte
+   *
+   * @external https://docs.nervos.org/docs/essays/faq#how-do-you-calculate-transaction-fee
+   */
   const rawTransaction = ckb.generateRawTransaction({
     fromAddress: addresses.testnetAddress,
     toAddress,
@@ -86,12 +88,6 @@ const bootstrap = async () => {
     deps: ckb.config.secp256k1Dep,
   })
 
-  rawTransaction.witnesses = rawTransaction.inputs.map(() => '0x')
-  rawTransaction.witnesses[0] = {
-    lock: '',
-    inputType: '',
-    outputType: ''
-  }
 
   const signedTx = ckb.signTransaction(privateKey)(rawTransaction)
   /**