+++ packages/ckb-sdk-core/types/global.d.ts
@@ -6,12 +6,6 @@ interface DepCellInfo {
   depType: CKBComponents.DepType
 }
 
-interface CachedCell extends CKBComponents.CellIncludingOutPoint {
-  status: string
-  dataHash: string
-  type?: CKBComponents.Script | null
-}
-
 type StructuredWitness = CKBComponents.WitnessArgs | CKBComponents.Witness
 
 declare namespace LoadCellsParams {
@@ -38,14 +32,34 @@ declare namespace RawTransactionParams {
   type LockHash = string
   type PublicKeyHash = string
   type Capacity = string | bigint
-  export type Cell = Pick<CachedCell, 'dataHash' | 'type' | 'capacity' | 'outPoint'>
+  type Cell = {
+    data: string
+    lock: CKBComponents.Script
+    type?: CKBComponents.Script
+    capacity: CKBComponents.Capacity
+    outPoint: CKBComponents.OutPoint
+  }
+  type Fee =
+    | Capacity
+    | {
+        feeRate: Capacity
+        reconciler: (params: {
+          tx: CKBComponents.RawTransactionToSign
+          feeRate: Capacity
+          changeThreshold: Capacity
+          cells: Array<{ capacity: string; outPoint: CKBComponents.OutPoint }>
+          extraCount: number
+        }) => CKBComponents.RawTransactionToSign
+      }
   interface Base {
-    fee?: Capacity
+    fee?: Fee
     safeMode: boolean
     deps: DepCellInfo | DepCellInfo[]
     capacityThreshold?: Capacity
     changeThreshold?: Capacity
     changePublicKeyHash?: PublicKeyHash
+    witnesses?: Array<CKBComponents.WitnessArgs | CKBComponents.Witness>
+    outputsData?: Array<string>
   }
 
   interface Simple extends Base {