+++ packages/ckb-sdk-rpc/src/paramsFormatter.ts
@@ -16,6 +16,12 @@ const formatter = {
     }
     return format(arg)
   },
+  toArray: (format?: (args: any) => any) => (arg: any) => {
+    if (typeof format !== 'function' || !Array.isArray(arg)) {
+      return arg
+    }
+    return arg.map(format)
+  },
   toHash: (hash: string): RPC.Hash256 => {
     if (typeof hash !== 'string') {
       throw new StringHashTypeException(hash)
@@ -131,6 +137,15 @@ const formatter = {
   },
   toBoolean: (value: boolean) => {
     return !!value
+  },
+  toTransactionProof: (proof: CKBComponents.TransactionProof): RPC.TransactionProof => {
+    if (!proof) return proof
+    const { blockHash: block_hash, witnessesRoot: witnesses_root, ...rest } = proof
+    return {
+      block_hash,
+      witnesses_root,
+      ...rest,
+    }
   },
 }